Adjusted text color for label matcher (#7309)
